[rework] Split aparet probeConfig and Stats
